Add x86 runners to CI (#529)
Excludes mac as x86 mac is tier 3 for Julia.
Whether `Int64` was used or tests adapted to using `Int` was based on smallness of change.
Includes workaround for Issue #530.

+++ test/test_vscode_example_functions.jl
@@ -13,6 +13,6 @@ end
 
 # exact location in file matters
 function fibcall(n)
-    fib(Int64(n))
+    fib(Int(n))
     fib(n)
 end
